:sparkles: new: working create client

+++ src/main/java/com/evertonvsf/managementsystem/dao/crud/users/client/ClientCRUD.java
@@ -4,7 +4,7 @@ import com.evertonvsf.managementsystem.dao.crud.users.UserCRUD;
 import com.evertonvsf.managementsystem.models.users.Client;
 
 public interface ClientCRUD extends UserCRUD<Client> {
-    public Client findByCpf(Integer cpf);
-    public boolean deleteByCpf(Integer cpf);
+    public Client findByCpf(long cpf);
+    public boolean deleteByCpf(long cpf);
 
 }
